pyproject.toml
[tool.poetry]
name = "lntopo"
version = "0.1.0"
description = ""
authors = ["Christian Decker <decker.christian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "7.1.2"
pyln-proto = "^0.10"
networkx = "^2.5"
tqdm = "^4.50"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lntopo-cli = 'lntopo.__main__:cli'